Implement filter option for tasks
Implement a new filter option for the DSL tasks. The filter takes a regular
expression as parameter. If the filter is set items are only evaluated when
their names match the filter expression.

For example this can be used to update only jobs inside a specific folder:
gradlew dslUpdateJenkins --filter="^FolderName/.*"

Change-Id: Ifd9e92c1d356e5c9ad478343aa5a0859463cbc73