RestJobManagement: Upload items in defined order
So far the order in which items were uploaded to Jenkins was undefined,
it depended on the order in which the DSL files were loaded. As a result
the upload of a job could fail because the parent folder was not created
yet.

To fix this the RestJobManagement now parses all DSL scripts first, and
then uploads the changes in this order: First all folders, then all
jobs, and finally all views.

Change-Id: I3f3af6f129800683dbb309f92a086d5fd92475fd