Add new JobBuilder implementation and PipelineBuilder
The PipelineBuilder adds support to easily create multiple versions of
the same pipeline, e.g. for different branches. For this a new
implementation of the JobBuilder class was needed which will replace the
old JobBuilder with the next major release. Until then it is called
JobBuilder2. The old JobBuilder has been marked as deprecated.

Change-Id: Ia83cde16cf28eae6d826b08721e3570f91ffa902