Resolve #25: Use localGroovy() instead of hardcoding the Groovy version
This was already tried in 17bdf9a before and reverted in afb98ca because
all functional tests failed. The problem was that the build files for
the tests did not add a localGroovy() dependency. But since localGroovy()
does not declare a real dependency and instead only adds the local Groovy
to the classpath all users of the plugin (and also the tests) have to add
localGroovy() to their dependencies.