Fix #12: Add support for auto-generated DSL and extensions
Add support for the auto-generated DSL [1] and Job DSL extensions [2].

This is implemented using the Jenkins test-harness to execute the Job DSL
scripts in a temporary Jenkins instance. This way the discovery of plugin
extension points works the same way as it does in a Job DSL seed job.

The list of plugins can be configured in the jenkinsPlugins configuration
in Gradle.

Also change the next version to 3.0.0 because this is a major change.

[1] https://github.com/jenkinsci/job-dsl-plugin/wiki/Automatically-Generated-DSL
[2] https://github.com/jenkinsci/job-dsl-plugin/wiki/Extending-the-DSL