Fix #65: Do not add common DSL closure multiple times in PipelineBuilder
Pass the common DSL closure to the build method of the job builder instead
of adding it to the list of dsl closures. This prevents it from being added
multiple times when PipelineBuilder.build() is called multiple times.

Add byte-buddy to the test dependencies to enable Spock to mock
PipelineJobBuilder which is not an interface [1][2].

[1] http://spockframework.org/spock/docs/1.1/all_in_one.html#_mocking_classes
[2] http://spockframework.org/spock/docs/1.1/all_in_one.html#_1_1_rc_3_released_2016_10_17