Make Groovy Seed Job more resilient to errors when creating/updating jobs or views
This might be a edge-case situation but in the case of #79, when there's a bug
in the seed job groovy script that produces an error before the seed job being
able to update itself a manual intervention is required. With this change,
errors while processing the xml would be ignored and the seed job would be
able to update itself if none of the xml errors are related to it.